Example 3.2 let us C
In a company an employee is paid as under:
if his basic salary is less than Rs. 1500, then HRA=10% of basic salary and DA=90% of basic salary. If his salary is either equal to or above Rs. 1500, then HRA =Rs,500 and DA=98% of basic salary. if the employee's salary is input through the keyboard write a program to find his gross salary.